+++ R/estimate_means.R
@@ -44,19 +44,34 @@
 #' @param estimate Character string, indicating the type of target population
 #' predictions refer to. This dictates how the predictions are "averaged" over
 #' the non-focal predictors, i.e. those variables that are not specified in
-#' `by` or `contrast`.
-#' - `"average"` (default): Takes the mean value for non-focal numeric
-#'   predictors and marginalizes over the factor levels of non-focal terms,
-#'   which computes a kind of "weighted average" for the values at which these
-#'   terms are hold constant. These predictions are a good representation of the
-#'   sample, because all possible values and levels of the non-focal predictors
-#'   are considered. It answers the question, "What is the predicted value for
-#'   an 'average' observation in *my data*?". Cum grano salis, it refers to
-#'   randomly picking a subject of your sample and the result you get on
-#'   average. This approach is the one taken by default in the `emmeans`
-#'   package.
-#' - `"population"`: Non-focal predictors are marginalized over the observations
-#'   in the sample, where the sample is replicated multiple times to produce
+#' `by` or `contrast`. We can roughly distinguish between "modelbased" and
+#' "empirical" predictions.
+#' - `"typical"` (default): Predictions are made for observations that are
+#'   represented by a data grid, which is built from all combinations of the
+#'   predictor levels in `by` (the focal terms). `"typical"` then takes the mean
+#'   value for non-focal numeric predictors and marginalizes over the factor
+#'   levels of non-focal terms, which computes a kind of "weighted average" for
+#'   the values at which these terms are hold constant. These predictions are
+#'   useful for comparing defined "groups" and are still a good representation
+#'   of the sample, because all possible values and levels of the non-focal
+#'   predictors are considered (averaged over). It answers the question, "What
+#'   would be the average outcome for a 'typical' observation?", where 'typical'
+#'   refers to subjects represented by (i.e., that share the characteristics
+#'   from) the data grid. This approach is the one taken by default in the
+#'   `emmeans` package.
+#' - `"average"`: Predictions are made for each observation in the sample. Then,
+#'   the average of all predictions is calculated within all groups (or levels)
+#'   of the focal terms defined in `by`. These predictions are the closest
+#'   representation of the sample, because `estimate = "average"` averages
+#'   across the full sample, where groups (in `by`) are not represented by a
+#'   balanced data grid, but rather the empirical distributions of the
+#'   characteristics of the sample. It answers the question, "What is the
+#'   predicted value for an average observation (from a certain group in `by`)
+#'   in my data?".
+#' - `"population"`: Each observation is "cloned" multiple times, where each
+#'   duplicate gets one of the levels from the focal terms in `by`. We then have
+#'   one "original" and several copies of that original, each varying in the levels
+#'   of the focal terms. Hence, the sample is replicated multiple times to produce
 #'   "counterfactuals" and then takes the average of these predicted values
 #'   (aggregated/grouped by the focal terms). It can be considered as
 #'   extrapolation to a hypothetical target population. Counterfactual
@@ -69,13 +84,24 @@
 #'
 #' In other words, the distinction between estimate types resides in whether
 #' the prediction are made for:
-#' - A specific "individual" from the sample (i.e., a specific combination of
-#'   predictor values): this is what is obtained when using [`estimate_relation()`]
-#'   and the other prediction functions.
-#' - An average individual from the sample: obtained with
-#'   `estimate_means(..., estimate = "average")`
-#' - The broader, hypothetical target population: obtained with
-#'   `estimate_means(..., estimate = "population")`
+#' - *modelbased predictions* (which are useful to look at differences between
+#'   typical groups, or for visualization)
+#'   - A specific individual from the sample (i.e., a specific combination of
+#'     predictor values for focal and non-focal terms): this is what is obtained
+#'     when using [`estimate_relation()`] and the other prediction functions.
+#'   - A typical individual from the sample: obtained with
+#'     `estimate_means(..., estimate = "typical")`
+#' - *empirical predictions* (which are useful if you want a realistic picture
+#'   of your sample, assuming that it is representative for a special population
+#'   (option `"average"`), or useful for "what-if" scenarios, especially if you
+#'   want to make unbiased comparisons (G-computation, option `"population"`))
+#'   - The average individual from the sample: obtained with
+#'     `estimate_means(..., estimate = "average")`
+#'   - The broader, hypothetical target population: obtained with
+#'     `estimate_means(..., estimate = "population")`
+#'
+#' You can set a default option for the `estimate` argument via `options()`,
+#' e.g. `options(modelbased_estimate = "average")`
 #' @param backend Whether to use `"emmeans"` or `"marginaleffects"` as a backend.
 #' Results are usually very similar. The major difference will be found for mixed
 #' models, where `backend = "marginaleffects"` will also average across random
@@ -114,6 +140,16 @@
 #' @inherit estimate_slopes details
 #'
 #' @return A data frame of estimated marginal means.
+#'
+#' @section Global Options to Customize Estimation of Marginal Means:
+#'
+#' - `modelbased_backend`: `options(modelbased_backend = <string>)` will set a
+#'   default value for the `backend` argument and can be used to set the package
+#'   used by default to calculate marginal means. Can be `"marginalmeans"` or
+#'   `"emmeans"`.
+#'
+#' - `modelbased_estimate`: `options(modelbased_estimate = <string>)` will
+#'   set a default value for the `estimate` argument.
 #'
 #' @references
 #' Dickerman, Barbra A., and Miguel A. Hernán. 2020. Counterfactual Prediction
@@ -180,7 +216,7 @@ estimate_means <- function(model,
                            by = "auto",
                            predict = NULL,
                            ci = 0.95,
-                           estimate = "average",
+                           estimate = getOption("modelbased_estimate", "typical"),
                            transform = NULL,
                            backend = getOption("modelbased_backend", "marginaleffects"),
                            verbose = TRUE,
@@ -188,7 +224,7 @@ estimate_means <- function(model,
   # validate input
   estimate <- insight::validate_argument(
     estimate,
-    c("average", "population", "specific")
+    c("typical", "population", "specific", "average")
   )
 
   if (backend == "emmeans") {
@@ -220,10 +256,12 @@ estimate_means <- function(model,
   info <- attributes(estimated)
 
   # Table formatting
-  attr(means, "table_title") <- c(ifelse(
-    estimate == "specific",
-    "Model-based Predictions",
-    "Estimated Marginal Means"
+  attr(means, "table_title") <- c(switch(
+    estimate,
+    specific = "Model-based Predictions",
+    typical = "Estimated Marginal Means",
+    average = "Average Predictions",
+    population = "Average Counterfactual Predictions"
   ), "blue")
   attr(means, "table_footer") <- .table_footer(
     means,